Adapt the test suite to scala 3.6
