--- tests/neg/i20415.scala
@@ -1,2 +1,0 @@
-class Foo:
-  given ord: Ordering[Int] = summon[Ordering[Int]] // error
