--- tests/neg/i6716.check
@@ -1,14 +1,0 @@
--- Error: tests/neg/i6716.scala:11:39 ----------------------------------------------------------------------------------
-11 |  given Monad[Bar] = summon[Monad[Foo]] // error
-   |                                       ^
-   |                             Result of implicit search for Monad[Foo] will change.
-   |                             Current result Bar.given_Monad_Bar will be no longer eligible
-   |                               because it is not defined before the search position.
-   |                             Result with new rules: Foo.given_Monad_Foo.
-   |                             To opt into the new rules, compile with `-source future` or use
-   |                             the `scala.language.future` language import.
-   |
-   |                             To fix the problem without the language import, you could try one of the following:
-   |                               - use a `given ... with` clause as the enclosing given,
-   |                               - rearrange definitions so that Bar.given_Monad_Bar comes earlier,
-   |                               - use an explicit argument.
