+++ tests/neg/i7294.check
@@ -1,25 +1,9 @@
--- Error: tests/neg/i7294.scala:7:10 -----------------------------------------------------------------------------------
-7 |  case x: T => x.g(10) // error // error
-  |          ^
-  |          Result of implicit search for scala.reflect.TypeTest[Nothing, T] will change.
-  |          Current result foo.f will be no longer eligible
-  |            because it is not defined before the search position.
-  |          Result with new rules: No Matching Implicit.
-  |          To opt into the new rules, compile with `-source future` or use
-  |          the `scala.language.future` language import.
-  |
-  |          To fix the problem without the language import, you could try one of the following:
-  |            - use a `given ... with` clause as the enclosing given,
-  |            - rearrange definitions so that foo.f comes earlier,
-  |            - use an explicit argument.
-  |
-  |          where:    T is a type in given instance f with bounds <: foo.Foo
--- [E007] Type Mismatch Error: tests/neg/i7294.scala:7:18 --------------------------------------------------------------
-7 |  case x: T => x.g(10) // error // error
-  |               ^^^^^^^
-  |               Found:    Any
-  |               Required: T
-  |
-  |               where:    T is a type in given instance f with bounds <: foo.Foo
+-- [E007] Type Mismatch Error: tests/neg/i7294.scala:7:15 --------------------------------------------------------------
+7 |  case x: T => x.g(10) // error
+  |               ^
+  |               Found:    (x : Nothing)
+  |               Required: ?{ g: ? }
+  |               Note that implicit conversions were not tried because the result of an implicit conversion
+  |               must be more specific than ?{ g: [applied to (10) returning T] }
   |
   | longer explanation available when compiling with `-explain`
