+++ tests/neg/i7294.scala
@@ -4,7 +4,7 @@ package foo
 trait Foo { def g(x: Any): Any }
 
 inline given f[T <: Foo]: T = ??? match {
-  case x: T => x.g(10) // error // error
+  case x: T => x.g(10) // error
 }
 
 @main def Test = f
