+++ tests/pos/i20415.scala
@@ -1,0 +1,2 @@
+class Foo:
+  given ord: Ordering[Int] = summon[Ordering[Int]]
