+++ tests/pos/looping-givens.scala
@@ -1,4 +1,3 @@
-import language.future
 
 class A
 class B
@@ -6,6 +5,6 @@ class B
 given joint(using a: A, b: B): (A & B) = ???
 
 def foo(using a: A, b: B) =
-  given aa: A = summon       // error
-  given bb: B = summon       // error
-  given ab: (A & B) = summon // error
+  given aa: A = summon       // resolves to a
+  given bb: B = summon       // resolves to b
+  given ab: (A & B) = summon // resolves to joint(aa, bb)
